requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
pymongo[srv]==3.12.0
flask-restx==1.3.0
Werkzeug==2.0.1
python-dotenv==1.0.0
requests